requirements.txt
numpy==1.19.3
Pillow
matplotlib